+++ mne/transforms.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # Authors: Alexandre Gramfort <alexandre.gramfort@telecom-paristech.fr>
 #          Christian Brodbeck <christianbrodbeck@nyu.edu>
 #
@@ -10,11 +11,12 @@ import numpy as np
 from numpy import sin, cos
 from scipy import linalg
 
+from .fixes import _get_sph_harm
 from .io.constants import FIFF
 from .io.open import fiff_open
 from .io.tag import read_tag
 from .io.write import start_file, end_file, write_coord_trans
-from .utils import check_fname, logger
+from .utils import check_fname, logger, verbose
 from .externals.six import string_types
 
 
@@ -463,7 +465,7 @@ def invert_transform(trans):
     return Transform(trans['to'], trans['from'], linalg.inv(trans['trans']))
 
 
-def transform_surface_to(surf, dest, trans):
+def transform_surface_to(surf, dest, trans, copy=False):
     """Transform surface to the desired coordinate system
 
     Parameters
@@ -475,11 +477,13 @@ def transform_surface_to(surf, dest, trans):
         FIFF types.
     trans : dict
         Transformation.
+    copy : bool
+        If False (default), operate in-place.
 
     Returns
     -------
     res : dict
-        Transformed source space. Data are modified in-place.
+        Transformed source space.
     """
     if isinstance(dest, string_types):
         if dest not in _str_to_frame:
@@ -548,37 +552,434 @@ def get_ras_to_neuromag_trans(nasion, lpa, rpa):
     return trans
 
 
-def _sphere_to_cartesian(theta, phi, r):
-    """Transform spherical coordinates to cartesian"""
-    z = r * np.sin(phi)
-    rcos_phi = r * np.cos(phi)
-    x = rcos_phi * np.cos(theta)
-    y = rcos_phi * np.sin(theta)
-    return x, y, z
+###############################################################################
+# Spherical coordinates and harmonics
+
+def _cart_to_sph(cart):
+    """Convert Cartesian coordinates to spherical coordinates.
+
+    Parameters
+    ----------
+    cart_pts : ndarray, shape (n_points, 3)
+        Array containing points in Cartesian coordinates (x, y, z)
+
+    Returns
+    -------
+    sph_pts : ndarray, shape (n_points, 3)
+        Array containing points in spherical coordinates (rad, azimuth, polar)
+    """
+    assert cart.ndim == 2 and cart.shape[1] == 3
+    cart = np.atleast_2d(cart)
+    out = np.empty((len(cart), 3))
+    out[:, 0] = np.sqrt(np.sum(cart * cart, axis=1))
+    out[:, 1] = np.arctan2(cart[:, 1], cart[:, 0])
+    out[:, 2] = np.arccos(cart[:, 2] / out[:, 0])
+    return out
 
 
-def _polar_to_cartesian(theta, r):
+def _sph_to_cart(sph):
+    """Convert spherical coordinates to Cartesion coordinates."""
+    assert sph.ndim == 2 and sph.shape[1] == 3
+    sph = np.atleast_2d(sph)
+    out = np.empty((len(sph), 3))
+    out[:, 2] = sph[:, 0] * np.cos(sph[:, 2])
+    xy = sph[:, 0] * np.sin(sph[:, 2])
+    out[:, 0] = xy * np.cos(sph[:, 1])
+    out[:, 1] = xy * np.sin(sph[:, 1])
+    return out
+
+
+def _get_n_moments(order):
+    """Compute the number of multipolar moments (spherical harmonics).
+
+    Equivalent to [1]_ Eq. 32.
+
+    .. note:: This count excludes ``degree=0`` (for ``order=0``).
+
+    Parameters
+    ----------
+    order : array-like
+        Expansion orders, often ``[int_order, ext_order]``.
+
+    Returns
+    -------
+    M : ndarray
+        Number of moments due to each order.
+    """
+    order = np.asarray(order, int)
+    return (order + 2) * order
+
+
+def _sph_to_cart_partials(az, pol, g_rad, g_az, g_pol):
+    """Convert spherical partial derivatives to cartesian coords.
+
+    Note: Because we are dealing with partial derivatives, this calculation is
+    not a static transformation. The transformation matrix itself is dependent
+    on azimuth and polar coord.
+
+    See the 'Spherical coordinate sytem' section here:
+    wikipedia.org/wiki/Vector_fields_in_cylindrical_and_spherical_coordinates
+
+    Parameters
+    ----------
+    az : ndarray, shape (n_points,)
+        Array containing spherical coordinates points (azimuth).
+    pol : ndarray, shape (n_points,)
+        Array containing spherical coordinates points (polar).
+    sph_grads : ndarray, shape (n_points, 3)
+        Array containing partial derivatives at each spherical coordinate
+        (radius, azimuth, polar).
+
+    Returns
+    -------
+    cart_grads : ndarray, shape (n_points, 3)
+        Array containing partial derivatives in Cartesian coordinates (x, y, z)
+    """
+    sph_grads = np.c_[g_rad, g_az, g_pol]
+    cart_grads = np.zeros_like(sph_grads)
+    c_as, s_as = np.cos(az), np.sin(az)
+    c_ps, s_ps = np.cos(pol), np.sin(pol)
+    trans = np.array([[c_as * s_ps, -s_as, c_as * c_ps],
+                      [s_as * s_ps, c_as, c_ps * s_as],
+                      [c_ps, np.zeros_like(c_as), -s_ps]])
+    cart_grads = np.einsum('ijk,kj->ki', trans, sph_grads)
+    return cart_grads
+
+
+def _deg_ord_idx(deg, order):
+    """Get the index into S_in or S_out given a degree and order."""
+    # The -1 here is because we typically exclude the degree=0 term
+    return deg * deg + deg + order - 1
+
+
+def _sh_negate(sh, order):
+    """Helper to get the negative spherical harmonic from a positive one"""
+    assert order >= 0
+    return sh.conj() * (-1. if order % 2 else 1.)  # == (-1) ** order
+
+
+def _sh_complex_to_real(sh, order):
+    """Helper function to convert complex to real basis functions.
+
+    Parameters
+    ----------
+    sh : array-like
+        Spherical harmonics. Must be from order >=0 even if negative orders
+        are used.
+    order : int
+        Order (usually 'm') of multipolar moment.
+
+    Returns
+    -------
+    real_sh : array-like
+        The real version of the spherical harmonics.
+
+    Notes
+    -----
+    This does not include the Condon-Shortely phase.
+    """
+
+    if order == 0:
+        return np.real(sh)
+    else:
+        return np.sqrt(2.) * (np.real if order > 0 else np.imag)(sh)
+
+
+def _sh_real_to_complex(shs, order):
+    """Convert real spherical harmonic pair to complex
+
+    Parameters
+    ----------
+    shs : ndarray, shape (2, ...)
+        The real spherical harmonics at ``[order, -order]``.
+    order : int
+        Order (usually 'm') of multipolar moment.
+
+    Returns
+    -------
+    sh : array-like, shape (...)
+        The complex version of the spherical harmonics.
+    """
+    if order == 0:
+        return shs[0]
+    else:
+        return (shs[0] + 1j * np.sign(order) * shs[1]) / np.sqrt(2.)
+
+
+def _compute_sph_harm(order, az, pol):
+    """Compute complex spherical harmonics of spherical coordinates."""
+    sph_harm = _get_sph_harm()
+    out = np.empty((len(az), _get_n_moments(order) + 1))
+    # _deg_ord_idx(0, 0) = -1 so we're actually okay to use it here
+    for degree in range(order + 1):
+        for order_ in range(degree + 1):
+            sph = sph_harm(order_, degree, az, pol)
+            out[:, _deg_ord_idx(degree, order_)] = \
+                _sh_complex_to_real(sph, order_)
+            if order_ > 0:
+                out[:, _deg_ord_idx(degree, -order_)] = \
+                    _sh_complex_to_real(_sh_negate(sph, order_), -order_)
+    return out
+
+
+###############################################################################
+# Thin-plate spline transformations
+
+# Adapted from code from the MATLAB file exchange:
+#    https://www.mathworks.com/matlabcentral/fileexchange/
+#            53867-3d-point-set-warping-by-thin-plate-rbf-function
+#    https://www.mathworks.com/matlabcentral/fileexchange/
+#            53828-rbf-or-thin-plate-splines-image-warping
+# Associated (BSD 2-clause) license:
+#
+# Copyright (c) 2015, Wang Lin
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met:
+#
+#     * Redistributions of source code must retain the above copyright
+#       notice, this list of conditions and the following disclaimer.
+#     * Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in
+#       the documentation and/or other materials provided with the distribution
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+class _TPSWarp(object):
+    """Transform points using thin-plate spline (TPS) warping
+
+    Notes
+    -----
+    Adapted from code by `Wang Lin <wanglin193@hotmail.com>`_.
+
+    References
+    ----------
+    .. [1] Bookstein, F. L. "Principal Warps: Thin Plate Splines and the
+           Decomposition of Deformations." IEEE Trans. Pattern Anal. Mach.
+           Intell. 11, 567-585, 1989.
+    """
+    def fit(self, source, destination, reg=1e-3):
+        from scipy.spatial.distance import cdist
+        assert source.shape[1] == destination.shape[1] == 3
+        assert source.shape[0] == destination.shape[0]
+        logger.info('Computing TPS warp coefficients using %d matched points'
+                    % (len(source),))
+        # Forward warping, different from image warping, use |dist|**2
+        dists = _tps(cdist(source, destination, 'sqeuclidean'))
+        # Y = L * w
+        # L: RBF matrix about source
+        # Y: Points matrix about destination
+        P = np.concatenate((np.ones((source.shape[0], 1)), source), axis=-1)
+        L = np.vstack([np.hstack([dists, P]),
+                       np.hstack([P.T, np.zeros((4, 4))])])
+        Y = np.concatenate((destination, np.zeros((4, 3))), axis=0)
+        # Regularize it a bit
+        L += reg * np.eye(L.shape[0])
+        self._destination = destination.copy()
+        self._weights = linalg.lstsq(L, Y)[0]
+        return self
+
+    @verbose
+    def transform(self, pts, verbose=None):
+        """Apply the warp.
+
+        Parameters
+        ----------
+        pts : shape (n_transform, 3)
+            Source points to warp to the destination.
+
+        Returns
+        -------
+        dest : shape (n_transform, 3)
+            The transformed points.
+        """
+        logger.info('Transforming %s points' % (len(pts),))
+        from scipy.spatial.distance import cdist
+        assert pts.shape[1] == 3
+        # for memory reasons, we should do this in ~100 MB chunks
+        out = np.zeros_like(pts)
+        n_splits = max(int((pts.shape[0] * self._destination.shape[0]) /
+                           (100e6 / 8.)), 1)
+        for this_out, this_pts in zip(np.array_split(out, n_splits),
+                                      np.array_split(pts, n_splits)):
+            dists = _tps(cdist(this_pts, self._destination, 'sqeuclidean'))
+            L = np.hstack((dists, np.ones((dists.shape[0], 1)), this_pts))
+            this_out[:] = np.dot(L, self._weights)
+        assert not (out == 0).any()
+        return out
+
+
+def _tps(distsq):
+    """Thin-plate function (r ** 2) * np.log(r)."""
+    # NOTE: For our warping functions, a radial basis like
+    # exp(-distsq / radius ** 2) could also be used
+    out = np.zeros_like(distsq)
+    mask = distsq > 0  # avoid log(0)
+    valid = distsq[mask]
+    out[mask] = valid * np.log(valid)
+    return out
+
+
+###############################################################################
+# Spherical harmonic approximation + TPS warp
+
+class SphericalSurfaceWarp(object):
+    """Warp surfaces via spherical harmonic smoothing and thin-plate splines.
+
+    Notes
+    -----
+    This class can be used to warp data from a source subject to
+    a destination subject, as described in [1]_. The procedure is:
+
+        1. Perform a spherical harmonic approximation to the source and
+           destination surfaces, which smooths them and allows arbitrary
+           interpolation.
+        2. Choose matched points on the two surfaces.
+        3. Use thin-plate spline warping (common in 2D image manipulation)
+           to generate transformation coefficients.
+        4. Warp points from the source subject (which should be inside the
+           original surface) to the destination subject.
+
+    .. versionadded:: 0.14
+
+    References
+    ----------
+    .. [1] Darvas F, Ermer JJ, Mosher JC, Leahy RM (2006). "Generic head
+           models for atlas-based EEG source analysis."
+           Human Brain Mapping 27:129–143
+    """
+    @verbose
+    def fit(self, source, destination, order=4, reg=1e-3, center=True,
+            verbose=None):
+        """Fit the warp from source points to destination points.
+
+        Parameters
+        ----------
+        source : array, shape (n_src, 3)
+            The source points.
+        destination : array, shape (n_dest, 3)
+            The destination points.
+        order : int
+            Order of the spherical harmonic fit.
+        reg : float
+            Regularization of the TPS warp.
+        center : bool
+            If True, center the points by fitting a sphere to points
+            that are in a reasonable region for head digitization.
+        verbose : bool, str, int, or None
+            If not None, override default verbose level (see mne.verbose)
+
+        Returns
+        -------
+        inst : instance of SphericalSurfaceWarp
+            The warping object (for chaining).
+        """
+        from .bem import _fit_sphere
+        logger.info('Computing TPS warp')
+        if center:
+            logger.info('    Centering data')
+            hsp = np.array([p for p in source
+                            if not (p[2] < -1e-6 and p[1] > 1e-6)])
+            src_center = _fit_sphere(hsp, disp=False)[1]
+            source = source - src_center
+            hsp = np.array([p for p in destination
+                            if not (p[2] < 0 and p[1] > 0)])
+            dest_center = _fit_sphere(hsp, disp=False)[1]
+            destination = destination - dest_center
+            logger.info('    Using centers %s -> %s'
+                        % (np.array_str(src_center, None, 3),
+                           np.array_str(dest_center, None, 3)))
+        assert source.shape[1] == destination.shape[1] == 3
+        self._destination = destination.copy()
+        # 1. Compute spherical coordinates of source and destination points
+        logger.info('    Converting to spherical coordinates')
+        src_rad_az_pol = _cart_to_sph(source).T
+        dest_rad_az_pol = _cart_to_sph(destination).T
+        # 2. Compute spherical harmonic coefficients for all points
+        logger.info('    Computing spherical harmonic approximation')
+        src_sph = _compute_sph_harm(order, *src_rad_az_pol[1:])
+        dest_sph = _compute_sph_harm(order, *dest_rad_az_pol[1:])
+        # 3. Fit spherical harmonics to both surfaces to smooth them
+        src_coeffs = linalg.lstsq(src_sph, src_rad_az_pol[0])[0]
+        dest_coeffs = linalg.lstsq(dest_sph, dest_rad_az_pol[0])[0]
+        # 4. Smooth both surfaces using these coefficients, and evaluate at
+        #     whichever has fewer points
+        if src_sph.shape[0] < dest_sph.shape[0]:
+            use_sph = src_sph
+            dest_rad_az_pol = src_rad_az_pol.copy()
+        else:
+            use_sph = dest_sph
+            src_rad_az_pol = dest_rad_az_pol.copy()
+        logger.info('    Matching %d points on smoothed surfaces'
+                    % (len(use_sph),))
+        src_rad_az_pol[0] = np.abs(np.dot(use_sph, src_coeffs))
+        dest_rad_az_pol[0] = np.abs(np.dot(use_sph, dest_coeffs))
+        # 5. Convert matched points to Cartesion coordinates and put back
+        source = _sph_to_cart(src_rad_az_pol.T)
+        source += src_center
+        destination = _sph_to_cart(dest_rad_az_pol.T)
+        destination += dest_center
+        # 6. Compute TPS warp of matched points from smoothed surfaces
+        self._warp = _TPSWarp().fit(source, destination, reg)
+        self._matched = np.array([source, destination])
+        return self
+
+    @verbose
+    def transform(self, source, verbose=None):
+        """Transform arbitrary source points to the destination.
+
+        Parameters
+        ----------
+        source : ndarray, shape (n_pts, 3)
+            Source points to transform. They do not need to be the same
+            points that were used to generate the model, although ideally
+            they will be inside the convex hull formed by the original
+            source points.
+        verbose : bool, str, int, or None
+            If not None, override default verbose level (see mne.verbose)
+
+        Returns
+        -------
+        destination : ndarray, shape (n_pts, 3)
+            The points transformed to the destination space.
+        """
+        return self._warp.transform(source)
+
+
+###############################################################################
+# Other transforms
+
+def _pol_to_cart(pol):
     """Transform polar coordinates to cartesian"""
-    x = r * np.cos(theta)
-    y = r * np.sin(theta)
-    return x, y
+    out = np.empty_like(pol)
+    out[:, 0] = pol[:, 0] * np.cos(pol[:, 1])
+    out[:, 1] = pol[:, 0] * np.sin(pol[:, 1])
+    return out
 
 
-def _cartesian_to_sphere(x, y, z):
-    """Transform cartesian coordinates to spherical"""
-    hypotxy = np.hypot(x, y)
-    r = np.hypot(hypotxy, z)
-    elev = np.arctan2(z, hypotxy)
-    az = np.arctan2(y, x)
-    return az, elev, r
+def _topo_to_sph(topo):
+    """Convert 2D topo coordinates to spherical coordinates."""
+    assert topo.ndim == 2 and topo.shape[1] == 2
+    sph = np.ones((len(topo), 3))
+    sph[:, 1] = -np.deg2rad(topo[:, 0])
+    sph[:, 2] = np.pi * topo[:, 1]
+    return sph
 
 
-def _topo_to_sphere(theta, radius):
-    """Convert 2D topo coordinates to spherical."""
-    sph_phi = (0.5 - radius) * 180
-    sph_theta = -theta
-    return sph_phi, sph_theta
-
+###############################################################################
+# Quaternions
 
 def quat_to_rot(quat):
     """Convert a set of quaternions to rotations